Museum of Art Update
I added a background color instead of a picture or gif since the photos in the slide needs to be seen more. I used cards in Bootstrap for the photos and captions used. I also added a navbar for the viewers to navigate from one museum to another. Downside is I am trying to figure out how to set bootstrap cars side by side.